refactor(domain): 관련 도메인 패키지화
