+++ src/main/java/calculator/service/StringSplitter.java
@@ -1,6 +1,6 @@
 package calculator.service;
 
-import calculator.domain.Delimiters;
+import calculator.domain.delimeter.Delimiters;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -15,7 +15,7 @@ public class StringSplitter {
         if (input == null || input.isEmpty()) {
             return new ArrayList<>();
         }
-        
+
         return List.of(input.split(delimiters.createRegexPattern()));
     }
 }
